added tests for implicit schur, minor changes
